Remove unused objects
