+++ app/controllers/tc/TaxCreditCalculatorController.scala
@@ -26,7 +26,6 @@ import play.api.libs.json._
 import play.api.mvc.Action
 import service.AuditEvents
 import uk.gov.hmrc.play.microservice.controller.BaseController
-
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
