+++ test/models/output/ESCCalculationSpec.scala
@@ -14,30 +14,17 @@
  * limitations under the License.
  */
 
-import models.output.OutputAPIModel.{AwardPeriod, Response}
-import models.output.esc.{Income, Savings, TaxAndNI}
+package models.output
+
+import models.output.esc.{TaxAndNI, Savings, Income}
 import org.joda.time.LocalDate
 import org.joda.time.format.DateTimeFormat
 import play.api.libs.json.Json
 import uk.gov.hmrc.play.test.UnitSpec
-import utils.{FakeCCCalculatorApplication, CCJsonLogger, Periods}
+import utils.{FakeCCCalculatorApplication, Periods}
 
-class OutputAPIModelSpec extends UnitSpec with FakeCCCalculatorApplication with CCJsonLogger {
 
-  "APIModelSpec" should {
-
-    "Response model's write to json" in {
-      val response = Response(AwardPeriod())
-      val js = Json.toJson(response)
-      js shouldBe Json.parse("""{"awardPeriod":{"tc":null,"tfc":null,"esc":null}}""")
-    }
-
-    "Total Award model's write to json" in {
-      val response = AwardPeriod()
-      val js = Json.toJson[AwardPeriod](response)
-      js shouldBe Json.parse("""{"tc":null,"tfc":null,"esc":null}""")
-    }
-  }
+class ESCCalculationSpec extends UnitSpec with FakeCCCalculatorApplication {
 
   "ESCCalculation models" should {
 
@@ -133,78 +120,4 @@ class OutputAPIModelSpec extends UnitSpec with FakeCCCalculatorApplication with 
         """.stripMargin)
     }
   }
-
-  "TFCCalculation models" should {
-
-    "TFCCalculation should write to JSON" in {
-      val formatter = DateTimeFormat.forPattern("yyyy-MM-dd")
-      val from = LocalDate.parse ("2016-08-01",formatter)
-      val until = LocalDate.parse ("2016-11-01",formatter)
-      val numberOfPeriods = 0
-
-      val contribution = models.output.tfc.Contribution(
-        BigDecimal(0.00),
-        BigDecimal(0.00),
-        BigDecimal(0.00)
-      )
-
-      val outputChild = models.output.tfc.OutputChild(
-        childCareCost =  BigDecimal(0.00),
-        childContribution = contribution
-      )
-
-      val tfcPeriod = models.output.tfc.TFCPeriod(
-        from = from,
-        until = until,
-        periodContribution = contribution,
-        children = List(outputChild)
-      )
-
-      val tfcCalculation = models.output.tfc.TFCCalculation(
-        from = from,
-        until = until,
-        householdContribution = contribution,
-        numberOfPeriods = numberOfPeriods.toShort,
-        periods = List(tfcPeriod)
-      )
-
-      val json = Json.toJson[models.output.tfc.TFCCalculation](tfcCalculation)
-
-      json shouldBe Json.parse(
-        """
-          |{
-          |	  "from": "2016-08-01",
-          |	  "until": "2016-11-01",
-          |   "householdContribution": {
-          |			  "parent": 0.0,
-          |			  "government": 0.0,
-          |			  "totalChildCareSpend": 0.0
-          |		},
-          |   "numberOfPeriods": 0,
-          |	  "periods": [
-          |   {
-          |		  "from": "2016-08-01",
-          |		  "until": "2016-11-01",
-          |		  "periodContribution": {
-          |			  "parent": 0.0,
-          |			  "government": 0.0,
-          |			  "totalChildCareSpend": 0.0
-          |		},
-          |		"children": [
-          |   {
-          |			  "childCareCost": 0.0,
-          |			  "childContribution": {
-          |				"parent": 0.0,
-          |				"government": 0.0,
-          |				"totalChildCareSpend": 0.0
-          |			}
-          |		}
-          |  ]
-          |	}
-          | ]
-          |}
-        """.stripMargin)
-    }
-  }
-
 }
